refactored upload download logic
