+++ web_hosting_manager/src/app.js
@@ -5,7 +5,7 @@ import { AppContainer } from 'react-hot-loader';
 import { loadTranslations, setLocale, syncTranslationWithStore } from 'react-redux-i18n';
 
 import Root from './containers/root';
-import { configureStore, history } from './store/configureStore';
+import { configureStore, history } from './store/configure_store';
 import loadLocale from './locales/loader';
 import { initTempFolder } from './safenet_comm/temp';
