--- web_hosting_manager/src/safenet_comm/helpers/dir_status.js
@@ -1,7 +1,0 @@
-export default class DirStats {
-  constructor() {
-    this.size = 0;
-    this.files = 0;
-    this.directories = 0;
-  }
-}
