--- web_hosting_manager/src/store/configureStore.js
@@ -1,6 +1,0 @@
-// @flow
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./configureStore.prod'); // eslint-disable-line global-require
-} else {
-  module.exports = require('./configureStore.dev'); // eslint-disable-line global-require
-}
