+++ web_hosting_manager/src/store/configure_store.dev.js
