+++ web_hosting_manager/src/store/configure_store.js
@@ -1,0 +1,6 @@
+// @flow
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./configure_store.prod'); // eslint-disable-line global-require
+} else {
+  module.exports = require('./configure_store.dev'); // eslint-disable-line global-require
+}
